Fix signal icon is not shown
The flag GSM of signal strength is set as false under OOS, if lower
layer does not notify signal strength again after in-service, the
signal level is based on CDMA, so that signal icon is unable to be
shown.

To fix it, change the preset flag GSM based on phone type.

Change-Id: I98a84f8701cbd4da9f9896d400048db733a224de
Signed-off-by: VenkatVishalV <venkatvishal124@gmail.com>
Signed-off-by: sayan7848 <sayan7848@gmail.com>